socialbarometer/pyproject.toml
[tool.poetry]
name = "socialbarometer"
version = "0.1.0"
description = "A social media sentiment analysis tool"
authors = ["Kuldeep Pal <kuldeep27396@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tweepy = "^4.10.0"
praw = "^7.6.0"
textblob = "^0.17.1"
asyncpg = "^0.25.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true